Move examples outside of ros_lib and add makefiles for gcc4mbed
