+++ rosserial_mbed/src/examples/ServiceClient/makefile
@@ -1,0 +1,8 @@
+PROJECT         := rosserial_mbed_ServiceClient
+DEVICES         := LPC1768
+#GCC4MBED_DIR    := /home/gary/devel/mbed/gcc4mbed
+INCDIRS         := ros_lib/
+NO_FLOAT_SCANF  := 1
+NO_FLOAT_PRINTF := 1
+
+include $(GCC4MBED_DIR)/build/gcc4mbed.mk
